% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bary.R
\name{bary}
\alias{bary}
\title{barycentric coordinates}
\usage{
bary(v1.ba, v2.ba, v3.ba, x.ba, y.ba)
}
\arguments{
\item{v1.ba}{the location of the first node in the target triangle}

\item{v2.ba}{the location of the second node in the target triangle}

\item{v3.ba}{the location of the third node in the target triangle}

\item{x.ba}{the x coordinate of the evaluated points}

\item{y.ba}{the y coordinate of the evaluated points}
}
\description{
obtain the barycentric coordinates
}
