% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beval.R
\name{beval}
\alias{beval}
\title{bivariate spline basis functions}
\usage{
beval(v.se, t.se, d.se, x.se, y.se)
}
\arguments{
\item{v.se}{the N*2 vertices matrix. Each row represents the location of
one vertex}

\item{t.se}{the n*3 triangle indices. Each row represents the number of
three vertices that forms a triangle. the order of vertices should be
counterclockwise}

\item{d.se}{degree of polynomial functions}

\item{x.se}{the x coordinate of evaluated points for basis functions}

\item{y.se}{the y coordinate of evaluated points for basis functions}
}
\description{
create bivariate spline basis functions on triangulation
}
