% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emp.area.R
\name{emp.area}
\alias{emp.area}
\title{the area of the domain}
\usage{
emp.area(fsb, n.area)
}
\arguments{
\item{fsb}{the list of the boundary points of the domain, where the}

\item{n.area}{the number of Monte Carlo samples}
}
\description{
calculate the area of the domain using Monte Carlo method
}
