% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/energy.R
\name{energy}
\alias{energy}
\title{energy matrix}
\usage{
energy(V.e, T.e, d.e, index.e)
}
\arguments{
\item{V.e}{vectices coordinates}

\item{T.e}{indices of the vectices of each triangle}

\item{d.e}{degree of barycentric polynomial}

\item{index.e}{=1, if the intergral term exists interaction of two coordinates; =2, otherwise;
default = 1}
}
\value{
penalty matrix with respect to the bivariate spline basis functions
}
\description{
create the penalty matrix with respect to bivariate spline basis functions
}
