% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loceval.R
\name{loceval}
\alias{loceval}
\title{local evaluation}
\usage{
loceval(lam1.le, lam2.le, lam3.le, bcoef.le)
}
\arguments{
\item{lam1.le}{the first coefficient of barycentric polynomial}

\item{lam2.le}{the second coefficient of barycentric polynomial}

\item{lam3.le}{the third coefficient of bary centric polynomial}
}
\description{
used in the major beval function to obtain the local evaluation
}
