% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothness.R
\name{smoothness}
\alias{smoothness}
\title{smoothness function}
\usage{
smoothness(V.sm, T.sm, d.sm, r.sm)
}
\arguments{
\item{V.sm}{vectices coordinates}

\item{T.sm}{indices of the vectices of each triangle}

\item{d.sm}{degree of barycentric polynomial}

\item{r.sm}{degree of derivative smoothness condition}
}
\value{
smoothness matrix
}
\description{
create the smoothness matrix
}
